Ming (#49)
* delete table_api

* added control plane and data plane

* load deployment on startup

* fix test

+++ crates/db/lib.rs
@@ -26,7 +26,7 @@ pub trait Connection {
 
 pub async fn get_conn(
     env_id: &str,
-    deploy_seq: i64,
+    deploy_seq: i32,
 ) -> Result<Rc<RefCell<dyn Connection>>> {
     // todo: use per project_id cache for connection pool.
     // let pool = mysql_simple::MySqlPool::new(
